pyproject.toml
[project]
name = "jack-bot"
version = "0.1.0"
description = "Telegram bot for Forest knowledge base"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21,<22",
    "python-dotenv>=1",
]

[project.scripts]
jack-bot = "bot.__main__:main"
